refactor: move core types to specific file
